Move on-demand zero-lemma checking into context_mcsat_relax.c
Split the mid-search checking logic (mcsat_relax_check_zero_lemma,
context_check_mcsat_relax_zero_lemmas) out of context.c into its own file,
following the existing context_solver.c/context_simplifier.c pattern of
sharing context_t internals via context_utils.h. Only the atom/monomial
construction logic (mcsat_relax_pprod and friends) stays in context.c, since
it's tightly coupled to the rest of the internalization pass there.

Required exposing three previously-static context.c helpers via
context_utils.h: context_mcsat_relaxation_enabled,
context_get_mcsat_relax_zero_lemma_done, translate_code_to_arith.
