+++ src/context/context.c
@@ -452,7 +452,7 @@ static void context_observe_mcsat_atom(context_t *ctx, term_t atom, literal_t l)
   }
 }
 
-static inline bool context_mcsat_relaxation_enabled(context_t *ctx) {
+bool context_mcsat_relaxation_enabled(context_t *ctx) {
   return ctx->mcsat_supplement && context_has_simplex_solver(ctx);
 }
 
@@ -484,7 +484,7 @@ static int_hmap_t *context_get_mcsat_relax_abstractions(context_t *ctx) {
   return map;
 }
 
-static int_hmap_t *context_get_mcsat_relax_zero_lemma_done(context_t *ctx) {
+int_hmap_t *context_get_mcsat_relax_zero_lemma_done(context_t *ctx) {
   int_hmap_t *map;
 
   map = ctx->mcsat_relax_zero_lemma_done;
@@ -3647,7 +3647,7 @@ static occ_t internalize_to_eterm(context_t *ctx, term_t t) {
  * - otherwise, x must be the code of an arithmetic variable v,
  *   we return v.
  */
-static thvar_t translate_code_to_arith(context_t *ctx, int32_t x) {
+thvar_t translate_code_to_arith(context_t *ctx, int32_t x) {
   eterm_t u;
   thvar_t v;
 
@@ -3803,157 +3803,6 @@ static thvar_t internalize_to_arith(context_t *ctx, term_t t) {
 }
 
 
-
-/****************************************************
- *  ON-DEMAND ZERO LEMMAS FOR THE MCSAT RELAXATION  *
- ***************************************************/
-
-/*
- * thvar_t for term t if t is already internalized to arithmetic, without
- * forcing internalization as a side effect of this (read-only) check.
- */
-static bool mcsat_relax_term_thvar(context_t *ctx, term_t t, thvar_t *x) {
-  term_t r;
-
-  r = intern_tbl_get_root(&ctx->intern, t);
-  if (!intern_tbl_root_is_mapped(&ctx->intern, r)) {
-    return false;
-  }
-  *x = translate_code_to_arith(ctx, intern_tbl_map_of_root(&ctx->intern, r));
-  return true;
-}
-
-/*
- * Look up (without building) the literal cached by mcsat_relax_cache_zero_atom
- * for the atom "t = 0". Read-only: never creates a new atom, which is unsafe
- * once the search is past the base level.
- */
-static bool mcsat_relax_cached_zero_atom(context_t *ctx, term_t t, literal_t *l) {
-  int_hmap_pair_t *r;
-
-  if (ctx->mcsat_relax_zero_atoms == NULL) {
-    return false;
-  }
-  r = int_hmap_find(ctx->mcsat_relax_zero_atoms, t);
-  if (r == NULL) {
-    return false;
-  }
-  *l = r->val;
-  return true;
-}
-
-/*
- * Check monomial (t, z) against the live simplex assignment and add at most
- * one zero-lemma clause for it:
- * - some factor xi = 0 while z != 0  ==>  add  not(xi=0) or (z=0)
- * - z = 0 while every factor is nonzero  ==>  add  not(z=0) or (x1=0) or ... or (xk=0)
- * Returns true if a clause was added.
- *
- * Only ever combines the literals cached by mcsat_relax_cache_zero_atom
- * (built eagerly at base level, see mcsat_relax_pprod) into new clauses;
- * it never builds a new atom itself, since that's unsafe mid-search.
- *
- * Those literals are typically fresh and unassigned, so nothing forces the
- * search to change the sampled assignment on its own: without the "done"
- * bitmask below, the same violation would be redetected and the same
- * clause re-added forever. At most one bit is ever set per call (bits
- * 0..n-1 for the forward clauses, bit 30 for the reverse clause), matching
- * the one-clause-per-monomial-per-call rule. Once every bit is set, this
- * monomial has nothing left to ever contribute and is skipped up front.
- */
-static bool mcsat_relax_check_zero_lemma(context_t *ctx, term_t t, term_t z, int_hmap_t *done_map) {
-  simplex_solver_t *simplex;
-  pprod_t *p;
-  thvar_t zx, fxi;
-  literal_t lz, *lf, *lits;
-  int_hmap_pair_t *done;
-  bool z_is_zero, some_factor_zero, added;
-  int32_t forward_i, full_mask;
-  uint32_t i, n;
-
-  p = pprod_term_desc(ctx->terms, t);
-  n = p->len;
-  assert(n < 30);
-  full_mask = ((1 << n) - 1) | (1 << 30);
-
-  done = int_hmap_get(done_map, t);
-  if (done->val < 0) {
-    done->val = 0;
-  }
-  if ((done->val & full_mask) == full_mask) {
-    return false;
-  }
-
-  if (!mcsat_relax_term_thvar(ctx, z, &zx) || !mcsat_relax_cached_zero_atom(ctx, z, &lz)) {
-    return false;
-  }
-  simplex = ctx->arith_solver;
-  z_is_zero = simplex_var_is_zero_in_assignment(simplex, zx);
-
-  lf = alloc_istack_array(&ctx->istack, n);
-  added = false;
-  forward_i = -1;
-  some_factor_zero = false;
-  for (i=0; i<n; i++) {
-    if (!mcsat_relax_term_thvar(ctx, p->prod[i].var, &fxi) ||
-        !mcsat_relax_cached_zero_atom(ctx, p->prod[i].var, lf+i)) {
-      goto cleanup;
-    }
-    if (simplex_var_is_zero_in_assignment(simplex, fxi)) {
-      if (!z_is_zero && forward_i < 0 && (done->val & (1 << i)) == 0) {
-        forward_i = (int32_t) i;
-      }
-      some_factor_zero = true;
-    }
-  }
-
-  if (forward_i >= 0) {
-    add_binary_clause(ctx->core, not(lf[forward_i]), lz);
-    done->val |= (1 << forward_i);
-    added = true;
-  } else if (z_is_zero && !some_factor_zero && (done->val & (1 << 30)) == 0) {
-    lits = alloc_istack_array(&ctx->istack, n+1);
-    lits[0] = not(lz);
-    for (i=0; i<n; i++) {
-      lits[i+1] = lf[i];
-    }
-    add_clause(ctx->core, n+1, lits);
-    free_istack_array(&ctx->istack, lits);
-    done->val |= (1 << 30);
-    added = true;
-  }
-
- cleanup:
-  free_istack_array(&ctx->istack, lf);
-  return added;
-}
-
-/*
- * Scan every tracked monomial abstraction once and add at most one zero-lemma
- * clause per monomial (batched across monomials, capped per monomial so the
- * per-round cost stays bounded as more lemma families are layered on later).
- * Returns true if at least one clause was added.
- */
-bool context_check_mcsat_relax_zero_lemmas(context_t *ctx) {
-  int_hmap_t *done_map;
-  int_hmap_pair_t *r;
-  bool added;
-
-  if (!context_mcsat_relaxation_enabled(ctx) || ctx->mcsat_relax_abstractions == NULL) {
-    return false;
-  }
-
-  done_map = context_get_mcsat_relax_zero_lemma_done(ctx);
-  added = false;
-  for (r = int_hmap_first_record(ctx->mcsat_relax_abstractions);
-       r != NULL;
-       r = int_hmap_next_record(ctx->mcsat_relax_abstractions, r)) {
-    if (term_kind(ctx->terms, r->key) == POWER_PRODUCT) {
-      added |= mcsat_relax_check_zero_lemma(ctx, r->key, r->val, done_map);
-    }
-  }
-  return added;
-}
 
 /*
  * Translate internalization code x to a bitvector variable
